🛠️refactor: useSegmentConvert 훅 분리, index.ts 수정
