+++ src/app/_components/client/Header/hooks/useHeaderLabel/useHeaderLabel.ts
@@ -1,0 +1,24 @@
+import { useMemo } from "react";
+
+interface useHeaderLabelProps {
+  segments: string[];
+}
+
+const useHeaderLabel = ({ segments }: useHeaderLabelProps) => {
+  return useMemo(() => {
+    let headerLabel = "";
+
+    switch (segments[0]) {
+      case "example":
+        headerLabel = "example";
+        break;
+
+      default:
+        break;
+    }
+
+    return { headerLabel };
+  }, [segments]);
+};
+
+export default useHeaderLabel;
