+++ src/app/_components/server/Icons/ArrowLeftIcon/ArrowLeftIcon.tsx
@@ -7,7 +7,7 @@ const ArrowLeftIcon = ({ size = 32, fill = "black", className = "", ...rest }: I
       height={size}
       fill={fill}
       className={className}
-      viewBox="0 0 10 16"
+      viewBox="0 0 10 20"
       xmlns="http://www.w3.org/2000/svg"
       {...rest}
     >
