+++ src/app/_components/server/icons/index.ts
@@ -1,2 +1,3 @@
 export { default as CloseIcon } from "./CloseIcon/CloseIcon";
 export { default as QuestionIcon } from "./QuestionIcon/QuestionIcon";
+export { default as ArrowLeftIcon } from "./ArrowLeftIcon/ArrowLeftIcon";
